Fixes for PR #8
